简化函数，复用代码，增加处理库
